v0.1.14.3 Created RubyMine.cs, added inher-ce from IselectUnit to Unit.cs
Try change gitignore
