+++ .gitignore
@@ -36,3 +36,4 @@ sysinfo.txt
 # Builds
 *.apk
 *.unitypackage
+*.log
