+++ Assets/Prefabs/Race/Human/Buildings/Barrack Middle.prefab
@@ -57808,12 +57808,20 @@ PrefabInstance:
       value: 3
       objectReference: {fileID: 0}
     - target: {fileID: 8001103419260674826, guid: b77eb349e03d76249affde7260657581, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8001103419260674826, guid: b77eb349e03d76249affde7260657581, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8001103419260674826, guid: b77eb349e03d76249affde7260657581, type: 3}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 8001103419260674826, guid: b77eb349e03d76249affde7260657581, type: 3}
       propertyPath: m_LocalPosition.y
-      value: 0
+      value: 2.1
       objectReference: {fileID: 0}
     - target: {fileID: 8001103419260674826, guid: b77eb349e03d76249affde7260657581, type: 3}
       propertyPath: m_LocalPosition.z
@@ -57846,6 +57854,10 @@ PrefabInstance:
     - target: {fileID: 8001103419260674826, guid: b77eb349e03d76249affde7260657581, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8001103419260674827, guid: b77eb349e03d76249affde7260657581, type: 3}
+      propertyPath: scaleX
+      value: 2
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: b77eb349e03d76249affde7260657581, type: 3}
