+++ Assets/Prefabs/Race/Human/Buildings/Ruby Mine.prefab
@@ -841,6 +841,17 @@ MonoBehaviour:
   changedPlayerMaterial: {fileID: 4697986118916888516}
   healthBar: {fileID: 8256937308369422717}
   bodyRadius: 0.7
+  miners:
+  - {fileID: 4860973939688638053}
+  - {fileID: 746960402892077069}
+  - {fileID: 7200886967852190680}
+  - {fileID: 8144378456770553218}
+  - {fileID: 289731682849272877}
+  - {fileID: 514235644358915061}
+  - {fileID: 9221849209405163777}
+  - {fileID: 1804440447031261747}
+  - {fileID: 3859688277655960190}
+  - {fileID: 6429510797522729475}
 --- !u!65 &5704448740866142751
 BoxCollider:
   m_ObjectHideFlags: 0
