--- Assets/Scripts/Buildings/GoldMine.cs
@@ -1,3 +1,0 @@
-public class GoldMine : Building, IBuilding
-{
-}
