+++ Assets/Scripts/Buildings/RubyMine.cs.meta
